pyproject.toml
[project]
name = "gen-text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.102.0",
    "python-dotenv>=1.1.1",
]
